requirements.txt
absl-py==0.9.0
asgiref==3.2.3
astor==0.8.1
attrs==19.3.0
blueprint==3.4.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
dill==0.3.1.1
Django==3.0.3
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
googleapis-common-protos==1.51.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.0.0
pkg-resources==0.0.0
promise==2.3
protobuf==3.11.3
pydot==1.4.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
tb-nightly==1.14.0a20190301
tensorboard==1.14.0
tensorflow==2.0.0a0
tensorflow-datasets==2.1.0
tensorflow-estimator==1.14.0
tensorflow-metadata==0.21.1
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
tqdm==4.43.0
urllib3==1.25.8
Werkzeug==0.16.1
wrapt==1.12.0
